+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 ### Fixed
 - Ignore unknown amino acids (X) for gravy score calculation
+- Save <NA> as pd.nan after clonal clustering in excel sheet
+- Convert V Identity to numeric, when importing for repertoire statistics
 
 ### Docs
 - Update instructions in `README.md` and documentation